// src/redux/tag-types.ts
export enum tagTypes {
    user = "user",
    admin = "admin",
    concern = "concern",
    inspectionGeneral = "inspectionGeneral",
    inspectionTire = "inspectionTire",
    contract = "contract",
    estimate = "estimate",
    appointment = "appointment",
    inspectionGroup = "inspectionGroup",
    // inspectionGroup = "inspectionGroup",
    // tags for services
    service = "service",
    singleService = "singleService",
    createService = "createService",
    searchService = "searchService",
    updateService = "updateService",
    deleteService = "deleteService",
    // tags for inspections
    inspection = "inspection",
    searchInspection = "searchInspection",
    getInspectionByCode = "getInspectionByCode",
    // tags for parts
    parts = "parts",
    createParts = "createParts",
    searchParts = "searchParts",
    deleteServiceLabour = "deleteServiceLabour",
    createSingleLabour = "createSingleLabour",
    createSinglePart = 'createSinglePart',
    updateSingleLabour = 'updateSingleLabour',
    deleteSinglePart = 'deleteSinglePart',
    deleteRelatedService = 'deleteRelatedService',
    assignRelatedService = 'assignRelatedService',
    updateRelatedService = 'updateRelatedService',
    deleteRelatedinspection = 'deleteRelatedinspection',
    updateRelatedInspection = 'updateRelatedInspection',
    assignRelatedInspection = 'assignRelatedInspection',
    getUnassignServiceInspectionAllData = 'getUnassignServiceInspectionAllData',
    getInspection = 'getInspection',
    createInspectionGroup = 'createInspectionGroup',
    assignConcernToService = 'assignConcernToService',
    getUnassignServiceInspectionAllDataService = 'getUnassignServiceInspectionAllDataService',
    serviceAdvisor = 'serviceAdvisor',
    technician = 'technician',
    createInstaspectionTask = "createInstaspectionTask",
    getInspectionData = 'getInspectionData',
    getSingleInspectionData = 'getSingleInspectionData',
    getSingleItem = "getSingleItem",
    getTireStatusTreadDepth = "getTireStatusTreadDepth",
    sendInvoiceToUser ='sendInvoiceToUser',
    sendReportToUser ='sendReportToUser'
}

export const tagTypesList = [
    tagTypes.user,
    tagTypes.admin,
    tagTypes.concern,
    tagTypes.inspectionGeneral,
    tagTypes.inspectionTire,
    tagTypes.service,
    tagTypes.contract,
    tagTypes.inspection,
    tagTypes.estimate,
    tagTypes.appointment,
    // tagTypes.inspectionGroup
    tagTypes.inspectionGroup,
    tagTypes.createService,
    // tagTypes.service,
    tagTypes.singleService,
    tagTypes.createService,
    tagTypes.searchService,
    tagTypes.updateService,
    tagTypes.deleteService,
    // tagTypes.inspection,
    tagTypes.searchInspection,
    tagTypes.getInspectionByCode,
    tagTypes.parts,
    tagTypes.createParts,
    tagTypes.searchParts,
    tagTypes.deleteServiceLabour,
    tagTypes.createSingleLabour,
    tagTypes.createSinglePart,
    tagTypes.updateSingleLabour,
    tagTypes.deleteSinglePart,
    tagTypes.deleteRelatedService,
    tagTypes.assignRelatedService,
    tagTypes.updateRelatedService,
    tagTypes.deleteRelatedinspection,
    tagTypes.updateRelatedInspection,
    tagTypes.assignRelatedInspection,
    tagTypes.getUnassignServiceInspectionAllData,
    tagTypes.getInspection,
    tagTypes.assignConcernToService,
    tagTypes.getUnassignServiceInspectionAllDataService,
    tagTypes.serviceAdvisor,
    tagTypes.technician,
    tagTypes.createInstaspectionTask,
    tagTypes.getInspectionData,
    tagTypes.getSingleInspectionData,
    tagTypes.getSingleItem,
    tagTypes.getTireStatusTreadDepth,
    tagTypes.sendInvoiceToUser,
    tagTypes.sendReportToUser
];